+++ core/src/main/java/com/graphhopper/routing/util/parsers/CarAccessParser.java
@@ -23,12 +23,7 @@ import com.graphhopper.routing.util.TransportationMode;
 import com.graphhopper.routing.util.WayAccess;
 import com.graphhopper.util.PMap;
 
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-
-import static java.util.Collections.emptyMap;
+import java.util.*;
 
 public class CarAccessParser extends AbstractAccessParser implements TagParser {
 
@@ -109,26 +104,26 @@ public class CarAccessParser extends AbstractAccessParser implements TagParser {
         if (way.hasTag("impassable", "yes") || way.hasTag("status", "impassable"))
             return WayAccess.CAN_SKIP;
 
-        // multiple restrictions needs special handling compared to foot and bike, see also motorcycle
+        // multiple restrictions needs special handling, see also motorcycle
+        boolean permittedWayConditionallyRestricted = getConditionalTagInspector().isPermittedWayConditionallyRestricted(way);
+        boolean restrictedWayConditionallyPermitted = getConditionalTagInspector().isRestrictedWayConditionallyPermitted(way);
         if (!firstValue.isEmpty()) {
             String[] restrict = firstValue.split(";");
-            boolean notConditionalyPermitted = !getConditionalTagInspector().isRestrictedWayConditionallyPermitted(way);
             for (String value : restrict) {
-                if (restrictedValues.contains(value) && notConditionalyPermitted)
+                if (restrictedValues.contains(value) && !restrictedWayConditionallyPermitted)
                     return WayAccess.CAN_SKIP;
-                if (intendedValues.contains(value))
+                if (intendedValues.contains(value) && !permittedWayConditionallyRestricted)
                     return WayAccess.WAY;
             }
         }
 
-        // do not drive street cars into fords
         if (isBlockFords() && ("ford".equals(highwayValue) || way.hasTag("ford")))
             return WayAccess.CAN_SKIP;
 
-        if (getConditionalTagInspector().isPermittedWayConditionallyRestricted(way))
+        if (permittedWayConditionallyRestricted)
             return WayAccess.CAN_SKIP;
-        else
-            return WayAccess.WAY;
+
+        return WayAccess.WAY;
     }
 
     @Override
@@ -154,8 +149,10 @@ public class CarAccessParser extends AbstractAccessParser implements TagParser {
             accessEnc.setBool(true, edgeId, intAccess, true);
         }
 
-        Map<String, Object> nodeTags = way.getTag("node_tags", emptyMap());
-        handleNodeTags(edgeId, intAccess, nodeTags);
+        if (way.hasTag("gh:barrier_edge")) {
+            List<Map<String, Object>> nodeTags = way.getTag("node_tags", Collections.emptyList());
+            handleBarrierEdge(edgeId, intAccess, nodeTags.get(0));
+        }
     }
 
     /**