+++ core/src/main/java/com/graphhopper/routing/util/parsers/FootAccessParser.java
@@ -23,10 +23,7 @@ import com.graphhopper.routing.util.TransportationMode;
 import com.graphhopper.routing.util.WayAccess;
 import com.graphhopper.util.PMap;
 
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 import static com.graphhopper.routing.ev.RouteNetwork.*;
 import static com.graphhopper.routing.util.PriorityCode.UNCHANGED;
@@ -132,13 +129,18 @@ public class FootAccessParser extends AbstractAccessParser implements TagParser 
         if (way.getTag("sac_scale") != null && !way.hasTag("sac_scale", allowedSacScale))
             return WayAccess.CAN_SKIP;
 
-        // no need to evaluate ferries or fords - already included here
-        if (way.hasTag("foot", intendedValues))
-            return WayAccess.WAY;
-
-        // check access restrictions
-        if (way.hasTag(restrictions, restrictedValues) && !getConditionalTagInspector().isRestrictedWayConditionallyPermitted(way))
-            return WayAccess.CAN_SKIP;
+        boolean permittedWayConditionallyRestricted = getConditionalTagInspector().isPermittedWayConditionallyRestricted(way);
+        boolean restrictedWayConditionallyPermitted = getConditionalTagInspector().isRestrictedWayConditionallyPermitted(way);
+        String firstValue = way.getFirstPriorityTag(restrictions);
+        if (!firstValue.isEmpty()) {
+            String[] restrict = firstValue.split(";");
+            for (String value : restrict) {
+                if (restrictedValues.contains(value) && !restrictedWayConditionallyPermitted)
+                    return WayAccess.CAN_SKIP;
+                if (intendedValues.contains(value) && !permittedWayConditionallyRestricted)
+                    return WayAccess.WAY;
+            }
+        }
 
         if (way.hasTag("sidewalk", sidewalkValues))
             return WayAccess.WAY;
@@ -149,11 +151,10 @@ public class FootAccessParser extends AbstractAccessParser implements TagParser 
         if (way.hasTag("motorroad", "yes"))
             return WayAccess.CAN_SKIP;
 
-        // do not get our feet wet, "yes" is already included above
-        if (isBlockFords() && (way.hasTag("highway", "ford") || way.hasTag("ford")))
+        if (isBlockFords() && ("ford".equals(highwayValue) || way.hasTag("ford")))
             return WayAccess.CAN_SKIP;
 
-        if (getConditionalTagInspector().isPermittedWayConditionallyRestricted(way))
+        if (permittedWayConditionallyRestricted)
             return WayAccess.CAN_SKIP;
 
         return WayAccess.WAY;
@@ -173,6 +174,11 @@ public class FootAccessParser extends AbstractAccessParser implements TagParser 
         } else {
             accessEnc.setBool(false, edgeId, intAccess, true);
             accessEnc.setBool(true, edgeId, intAccess, true);
+        }
+
+        if (way.hasTag("gh:barrier_edge")) {
+            List<Map<String, Object>> nodeTags = way.getTag("node_tags", Collections.emptyList());
+            handleBarrierEdge(edgeId, intAccess, nodeTags.get(0));
         }
     }
 }