+++ core/src/main/java/com/graphhopper/util/Constants.java
@@ -74,7 +74,7 @@ public class Constants {
     public static final int VERSION_NODE_CH = 0;
     public static final int VERSION_GEOMETRY = 6;
     public static final int VERSION_LOCATION_IDX = 5;
-    public static final int VERSION_EDGEKV_STORAGE = 2;
+    public static final int VERSION_KV_STORAGE = 2;
     /**
      * The version without the snapshot string
      */
@@ -147,7 +147,7 @@ public class Constants {
 
     public static String getVersions() {
         return VERSION_NODE + "," + VERSION_EDGE + "," + VERSION_GEOMETRY + "," + VERSION_LOCATION_IDX
-                + "," + VERSION_EDGEKV_STORAGE + "," + VERSION_SHORTCUT;
+                + "," + VERSION_KV_STORAGE + "," + VERSION_SHORTCUT;
     }
 
     public static String getMajorVersion() {